deliverable/backend/requirements.txt
# Core dependencies
flask==2.0.1
werkzeug==2.0.1
flask-cors==3.0.10
python-dotenv==0.19.0
gunicorn==20.1.0

# LLM and AI generation
litellm>=0.14.1
tenacity>=8.2.2
loguru>=0.7.0

# Utilities
requests==2.26.0
python-multipart==0.0.6
pathspec>=0.11.1
langchain_text_splitters>=0.0.1
chromadb>=0.4.13
